+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed issue with dependency graph failing to wait for step dependencies to
+  complete prior to performing more work
+
 ## 0.6.0 - 2020-04-21
 
 ### Changed